\docType{class}
\name{BMA-class}
\alias{BMA-class}
\alias{getBMA}
\alias{getBMA,BMA-method}
\alias{initialize,BMA-method}
\title{A Bayesian Model Averaging package}
\usage{
getBMA(object = "BMA")
}
\description{
Object of class \code{BMA} are created by the \code{fitBMA}
functions
}
\details{
An object of the class `BMA' has the following slots:
\itemize{ \item \code{coef} A matrix of coefficients
generated by all variable combinations \item \code{Rsquare}
A vector of R-square values from the regressions \item
\code{postProb} A vector of posterior probabilities \item
\code{postModel} A vector of posterior models \item
\code{postExp} A vector of posterior expectations for
coefficients \item \code{x} The first input: A matrix of
standardized covariates \item \code{y} The second input: A
numeric vector length of the number of regressions using
variable combinations }
}
\author{
Elif Ozdemir: \email{eozdemir@wustl.edu}
}

