\name{summaryBMA}
\alias{BMA-method}
\alias{summaryBMA}
\alias{summaryBMA,}
\title{Summarize BMA class objects}
\usage{
summaryBMA(object)
}
\arguments{
  \item{object}{A BMA class object}
}
\description{
Prints summary info for BMA class objects
}
\author{
Elif Ozdemir: \email{eozdemir@wustl.edu}
}
\seealso{
\code{\link{fitBMA}}

\code{\link{plotBMA}}
}

